requirements.txt
langchain==0.0.308
openai==0.27.6
tiktoken==0.4.0

python-dotenv==1.0.0
streamlit==1.27.0
streamlit-option-menu
